fixed checkstyle and jacoco exclusions

+++ pri/src/main/java/com/example/gradle/p1/exceptions/PrimeNotFoundException.java
@@ -2,7 +2,7 @@ package com.example.gradle.p1.exceptions;
 
 public class PrimeNotFoundException extends RuntimeException {
 
-    private final Long primeToSearchTo;
+    private final Integer primeToSearchTo;
 
     public PrimeNotFoundException(final Integer primeToSearchTo) {
         super("No Prime numbers found below " + primeToSearchTo);
